requirements.txt
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.1.0
scipy>=1.8.0
tensorflow-macos>=2.14.0
tensorflow-metal>=1.0.0
matplotlib>=3.6.0
pyEDFlib>=0.1.30
mne>=1.2.2
wget